Update patch set 10
Patch Set 10: Code-Review+2 Verified+1

Probably there are some small things that can be improved, but this looks really good as a first pass. Thanks again for doing this.

If there's nothing else you wanted to add (in this change as opposed to a later one), I'll go ahead and submit.

Patch-set: 10
Label: Code-Review=+2
Label: Verified=+1