Update patch set 1
Patch Set 1: Code-Review-1

(3 comments)

Please add some tests for this case. Yes, we have sucked so far at adding tests for LogServlet, but we're trying to do better. Look at RefServletTest for an example of how to set up a servlet and parse JSON in the test.

Patch-set: 1
Label: Code-Review=-1