ServletTest: Clean up buildJson methods
Put type argument first so the path and query string are adjacent,
which helps readability.

Change "additional" query string arg to just a query string. The
failure behavior if callers forget to start with "&" is confusing,
since it just screws up the format arg. Whether format is appended or
prepended is an implementation detail.

Change-Id: I9c2ac8dcd62c65bcffddcf6262114122e6049ee1