LogServlet: Default to follow=1
Although not the default in `git log`, there is a significant
usability improvement to defaulting to follow, particularly because we
don't expose a link the UI to turn this on.

There might still be performance concerns with this default, so allow
it to be configured in gitiles.config, using the same log.follow
option as in core Git.

Change-Id: I43fe926e0a22bf2f7e48cc5ab0194eb14d82fdd6