Improve styling/branding options
In addtion to the header introduce more customizable template parts
that are depending on 'customVariant' to make injection of html at
various place possible.

The 'customHeader' config variable is renamed to 'customVariant',
but the old name is still recognised as an alias for the new name to
ensure backwards compatibility.

The  new template parts are:
  customHeadTagPart  -> at the end of the <head> tag
  customFooter-> The footer tag with contents

The new fixed templates .footerFormatBadge and .footerPoweredBy can
be used in custom footers.

Change-Id: I95a97a3805f4fb3011bc932d4f9799ec5a2c78c4