Update patch set 7
Patch Set 7:

> > Now might be the time to consider releasing the gitiles
 > components
 > > on maven central instead of to the Google bucket.
 > 
 > I could try but I am not sure about how much effort this will be.
 > Would it be as simple as changing a few urls in BUCK and
 > fake_pom_deploy.xml?
 > 

It might be as simple as that.  On the other hand, it might not be :)

I'll have a look into it when I have time.  Probably not in the next week or so, though.

 > Since this change is holding 2.13 I wouldn't like to introduce even
 > more delay.

I agree.  Let's get it done so we can move on with 2.13.

Patch-set: 7
Reviewer: David Pursehouse <1011123@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Verified=0