RFC: Use google-java-format for styling
google-java-format is an open-source tool[1] for mechanically
converting Java code to follow the Google Java Style Guide[2]
formatting recommendations. The tool is inspired in part by gofmt[3],
which has been very successful in the Go community.

Any automatic formatter is going to make some decisions that not
everyone agrees with, but the goal is to avoid pointless discussions
in reviews that detract from actually understanding the code. My
personal experience with this tool is that it adds many more newlines
than I would write in my own code, leading to deeper indentation. It
took me a while to get used to it, but I'm fine with it now. (It's
worth pointing out that the deeper indentation is also more bearable
because Gitiles uses a 100-column limit; for an 80-column limit it
might be less readable.)

Converting all at once like this definitely hurts blame-ability, but
honestly at this point in Gitiles most of the lines were written once,
by me, so we're not losing too much.

Change-Id: I289c5ea97f416b9b709db1033bdb8357fdaa9690