Update patch set 1
Patch Set 1: Verified+1

Yea, this fixes an issue. The GitilesView was doing # -> %23 for us which was wrong. Downstream code to insert into HTML is happy to have one # in the URL as this is a safe way to reference an anchor.

Patch-set: 1
Reviewer: Shawn Pearce <1001240@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Verified=+1