Update patch set 5
Patch Set 5: Code-Review-1

(20 comments)

Thanks for this; I really appreciate the attention to detail with the backwards compatibility with any existing customHeader config as it helps us roll out this change with lower risk of breaking an existing override.

The approach is fine. All of my comments are style nits. I think if these are cleaned up we can just submit this.

Patch-set: 5
Label: Code-Review=-1