Markdown: support aside|note|promo blocks inside list items
This used to be accepted with the Pegdown parser, but was broken
during the switch to commonmark-java.

Also make the space between *** and the style optional, this was
accepted by the prior Pegdown parser and broke a few documents.
Its still unambiguous as BlockNoteExtension requires matching one
of three well known styles: "*** {aside|note|promo}".

Add a small unit test for the nested inside list behavior.

Change-Id: I34984ac14571e0b860b79096ee737a51cd4ea8ba