Update patch set 1
Patch Set 1:

> Need to check if bucklets allows subscription.  See issue 3311.
The issue is linked to the gitlies tracker,but  probably you ment the 

 > Need to check if bucklets allows subscription.  See issue 3311.

The issue link seems to be wrong (it links to gitiles on code.google.com, where the highest number seems to be 125). I guess it refers to a monorail bug.

Patch-set: 1
Reviewer: Gerrit User 1011624 <1011624@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Verified=0