WIP: Code Search UI
This patch aims at utilizing a zoekt server [1] as backend for
a code search in gitiles. A new view is added, /+search, that
displays the result of a code search in a given branch of a
repository for which the zoekt server can provide index data.
The view connects to the zoekt server that has been configured
in gitiles.conf in section [codeSearch], for example:

[codeSearch]
  zoektServer = http://zoekt.example.org:6070

Currently only the following URL schema is supported:
http://gitles.example.org:8080/<repo>/+search/<branch>?query=<query>

The links on the result page are already working, i.e. you can
directly jump to the file and even to the line of a match, but
there is not yet an input field in the UI for entering a query.

For testing in the DevServer, add the following line in
DefServer#defaultConfig():

    cfg.setString("codeSearch", null, "zoektServer",
       "http://<zoekt host>:6070");

[1] https://github.com/google/zoekt

Also-By: Saša Živkov <sasa.zivkov@sap.com>
Change-Id: I1a01de9ab60baa817da5acf398ca2b5201ccd342