WIP: Code Search UI
This patch aims at utilizing a zoekt server [1] as backend for
a code search in gitiles. A new view is added, /+search, that
displays the result of a code search in a given branch of a
repository for which the zoekt server can provide index data.

zoekt needs the name of a repository and at least one branch name
as input. Branch names are not easily available in gitiles, though.
Actually, only the repository index page calculates them, but the
other views treat branches just a refs, which are not usable as
input for zoekt yet. Therefore, the search form is visible only
on the repository index page, although for layout reasons it had
to be added to Common.soy. There is no possibility to select
multiple (or all) branches for a search query yet in the UI to
keep things simple.

The search form submits the query and branch to the search as:

   /<repo>/+search?branch=<branch>&query=<query>

The search view connects to the zoekt server, which has been
configured in gitiles.conf, section [codeSearch], for example:

[codeSearch]
  zoektServer = http://zoekt.example.org:6070

TODOs:
  - support searching multiple branches ("all branches" option in
    the dropdown list)
  - make the search available on views, where we can determine a
    branch from the view context
  - extract the HTTP communication code and imrprobe error handling
  - css

[1] https://github.com/google/zoekt

Also-By: Saša Živkov <sasa.zivkov@sap.com>
Change-Id: I1a01de9ab60baa817da5acf398ca2b5201ccd342