Update documentation to add --exclude option to eclipse script
Generating the eclipse project puts both servlet-api_2_5 and
servlet-api_3_0 in the classpath. We only need the latter in order
to run gitiles from eclipse. Having servlet-api_2_5 in the classpath
(in front of the servlet-api_3_0) produces errors and doesn't find
css styles.

The servlet-api_2_5 can be excluded by the --exclude option on the
eclipse script. Update the documentation to mention this.

Change-Id: I2174e46775519d7dc72422d69ebda9f0d8c1f1c3