Bazel: Generate eclipse project
servlet_api_2_5 can be excluded from the generated .classpath to avoid
classpath collisions between Servlet API 2.5 and Servlet API 3.0. To do
that, pass -x parameter to project.py invocation in project.sh:

  ... -n gitiles -r . -x servlet_api_2_5

TEST PLAN:

  $ tools/eclipse/project.sh

Open the project in Eclipse IDE and verify that it can be compiled and
the tests can be run.

Change-Id: If318321fb1a94888a9adea635458afee2b463682