Update patch set 1
Patch Set 1:

> tangent: we should drop buck for gitiles too.

There is series for Buck removal for Gitiles pending for review, that appears to be staled, because of inclusion of Eclipse project generation in Bazlets and referencing it from Gitiles, instead of adding it to Gitiles project itself.

In case it's not clear (or obvious) how open source works: contributors do the work when they have time for that. If you have different priorities and like to accelerate the process, by any means, feel free to help to address age comments and move things forward.

Patch-set: 1