Update patch set 8
Patch Set 8:

> Patch Set 8:
> 
> > Hi this is a big hack IMO.
> 
> Can you elaborate? What exactly is a big hack? The way how the python script location from bazlets  is discovered, or the project.py itself, the one that was added here: [1]?
>

Sorry I meant the way the python script location is discovered (catting an output of a txt file).

>  > Also the query based approach is really fragile, we deleted it from
>  > Bazel because it was always broken. Is there a reason not to use
>  > e4b (I know it is unstaffed) or even better IntelliJ?
> 
> We are open to investigate further options. However, to finalize the migration to Bazel and remove Buck in Gitiles and prolog-cafe, we must have 100% feature parity with Buck driven build. And that means fully support for Eclipse IDE.

I think e4b should do what you want I would be happy to have it work a bit better. Anyway this is a perfectly fine solution just one that is fragile. A better solution even for project generation would be to rely on aspect.

> 
> * [1]  https://gerrit-review.googlesource.com/95253

Patch-set: 8