Update patch set 8
Patch Set 8:

> > Patch Set 8:
 > >
 > > > Hi this is a big hack IMO.
 > >
 > > Can you elaborate? What exactly is a big hack? The way how the
 > python script location from bazlets  is discovered, or the
 > project.py itself, the one that was added here: [1]?
 > >
 > 
 > Sorry I meant the way the python script location is discovered
 > (catting an output of a txt file).

Can you offer a better solution, then writing a skylark rule, track down the py_binary and iterate over the transitive_sources and dump it in output?

Note, that it's not only restricted to Eclipse project generation (project.py). Another task would be to publish artifacts to local Maven repository or Google storage bucket. So we have mvn.py: [1] in bucklets and would like to do the same what we did manually in Buck with git submodules: abuse Bazel bazlets repository as transport layer. We cannot use bazel to actually publish with BUILD target, because we don't have stable input and output here (consider /me removing published artifacts in local or remote repository and repeating the buck build foo).

Note that I added mvn.py in follow up change in this series directly to this repository, but it was already noticed by reviewer. 

* [1] https://github.com/davido/bucklets/blob/master/tools/mvn.py

Patch-set: 8