Update patch set 8
Patch Set 8:

> Now most of the classpath entries are right. The remaining errors
 > are all about FakeHttpServletRequest missing some methods,
 > presumably because it's getting the Servlet API 3.0 instead of 2.5.

Yes exactly. This was the result of discussion with Sasa and Michael Ochmann during SAP Hackathon last year, how to make Gitiles project work in Eclipse, as there is a classpath collision between servlet API 2.5, needed for running tests and Servlet API 3.0 needed to run Gitiles outside of Google, where Web Server is not pinned to Servlet API 2.5, as it's the case say in Jetty 9. To recall, they were working on zoekt integration in Gitiles and were failing to make it work from Eclipse. As the result of this discussion, Sasa introdcued this change in Bucklets: [1] to allow to exclude arbitrary libraries during Eclipse classpath generation.

Note, that I'm not working on Gitiles, and don't really care. So I will hapily implement whatever you prefer here:

* A. Include both Servlet API 3.0 and Servlet API 2.5 in the .classpath: just drop -x parameter from project.sh
* B. Include Servlet API 2.5 and exclude Servlet API 3.0 from the .classpath instead: just replace the -x argument in project.sh
* C. Let it as is in the latest patch set

* [1] https://gerrit-review.googlesource.com/86327

Patch-set: 8
Label: Verified=0