Update patch set 12
Patch Set 12:

I just realised the version deployed on sonatype staging is showing version 1, which means it didn't take the version from the bzl file.  When I saw that before I assumed that's the version number you'd used.

I wonder if gs:// is failing because it doesn't like the version with -TEST suffix?

Patch-set: 12