Match GitHub TOC rendering for links
GitHub uses lowercase for anchor links, and correspondingly and TOC
references are written to use the same. This results in TOC's that work
when viewed in GitHub, but are broken in Gitiles, and vice-versa.

Switches to use lowercase for the name and id to ensure that it matches
the GitHub flavour of markdown.

Change-Id: I7a07430b183315e0d1433c4ff227a46e5c51bff3