Support type=RAW for files
Required when a URL should be resolved into a plain file, for example
configuration files. Extremely usable when markdown refer to a file to
be downloaded by the user.

Currently this is "hidden" feature as it is not presented to user within
the common template.

Configuration option gitiles.allowRaw controls this feature, default
disabled.

Signed-off-by: Alon Bar-Lev <alon.barlev@gmail.com>
Change-Id: I6f517049b795a83c5a577e1ba96f7fba1e6327c0