tools: run_dev: remove explicit sh
The dev script requires bash and indeed it has bash in its shebang.
Execute it via sh breaks it when default shell is not bash. Bazel
produce the dev script as executable so there is no need to wrap it in
explicit sh. Another option would be to replace sh with bash.

Signed-off-by: Alon Bar-Lev <alon.barlev@gmail.com>
Change-Id: Ie09ddf0cfec8495d8ec393a43ef7c385f43e969f