Update patch set 2
Patch Set 2:

> I do not understand how does this differ from any site that you have write access to.

The difference is, most sites that you have write access to, it's just *you* (or a small set of trusted users) that has write access. But say you're attaching Gitiles to a public Gerrit server that accepts contributions from any person on the internet. Now it's not just you that has write access, it's *anyone on the internet* that has write access.

> I see that you mainly concern about the content type mapping.

It's not about the mapping per se. The problem is that some, mainly older, browsers will completely ignore any and all hints you give it about the content type (Content-Disposition/Content-Type headers, etc.). In these browsers, if the code looks like HTML (or worse, like a jar), it will get rendered as HTML, and there is literally nothing the server can do to stop it.

Combine these:

* anyone on the internet can serve arbitrary data from your server, in your domain's security context
* you can't stop the arbitrary data form being interpreted as HTML/Java

and the risk of a malicious user publishing something that will steal your other users' data is too high.

Patch-set: 2
Reviewer: Gerrit User 1010008 <1010008@173816e5-2b9a-37c3-8a2e-48639d4f1153>