Add support for glue to permit some block HTML in Markdown
Allow a RootedDocServlet created by an integration application to
override the way MarkdownToHtml handles HtmlBlock AST nodes, casting
them into SafeHtml that can be appended directly to the HtmlBuilder.

This can allow an integrator to pass the user content through an HTML
sanitizer, and pass-through a subset it believes to be safe.

The default still drops everything on the floor, and it's not possible
to configure Gitiles to pass-through HTML by default.  Application
code changes are required to invoke the proper RootedDocServlet
constructor.

Change-Id: Id412bc8670d2c25f36c8486e0d3600d1cb9d0710