Add support for glue to permit some HTML in Markdown
Allow a RootedDocServlet subclass in an integration application to
override the way MarkdownToHtml handles HtmlBlock and HtmlInline AST
nodes, potentially casting them into SafeHtml which can be appended
to the HtmlBuilder.

This can allow an integrator to pass the user content through an HTML
scrubber and pass-through a subset their scrubber believes to be safe.

The default still drops everything on the floor, and its not possible
to configure Gitiles to pass-through HTML by default. Code changes are
required by subclassing RootedDocServlet.

Change-Id: Id412bc8670d2c25f36c8486e0d3600d1cb9d0710