Stamp final artifacts
Merge Implementation-Version into the final artifacts. This has two
advantages: for one, it can be later tracked down, from what release
this artifact was consumed; for another, in case a release was produced
without touching any source files but only bumping some dependencies,
that are not shaded in the final artifacts and thus do not contribute
to the SHA1 update of the final artifacts, the SHA1 would still change,
because of the changed content of the META-INF/MANIFEST.MF file. For
example building from this commit will produce the following manifest
content:

  $ bazel build java/com/google/gitiles/blame/cache:cache-stamped
  bazel-genfiles/java/com/google/gitiles/blame/cache/cache-stamped.jar

Unzipping the content of the cache-stamped.jar produce the following
implementation version:

  $ cat META-INF/MANIFEST.MF  | grep -i Impl
  Implementation-Version: v0.2-3-8-g55d6911-dirty

Change-Id: Idb6e7c635ae254188e1ee6592e0e0502c7c338a6