Change all foreach loops in Soy templates to use the for loop syntax
Soy supports 2 kinds of loops:

 foreach- for iterating over items in a collection
          e.g., {foreach $item in $list}...{/foreach}
 for- for indexed iteration
      e.g., {for $i in range(0, 10)}...{/for}

The reason Soy has 2 different loops is an accident of history.  Soy
didn't used to have a proper grammar for expressions and so the
alternate 'for...range' syntax was added to make it possible to write
indexed loops.  As the grammar has improved, having the two syntaxes
is no longer necessary and so we are eliminating one of them.

As of [1], the two forms are actually aliases for one another, so the
only difference is the keyword ('for' vs 'foreach').  While the
foreach loop is more popular, the 'for' terminology is more standard
so switch everything to that.

DO NOT SUBMIT until //lib/soy is updated to a version including [1].

[1] https://github.com/google/closure-templates/commit/4a7373333fee6c22784b48e29825b9cea3ffaae7

PiperOrigin-RevId: 179944099
Change-Id: I22fe13baffae5567d0e123395a62f2d837ea374c