Update patch set 3
Patch Set 3:

That is indeed a very good question here. This is actually a RFC. I did the same for Gerrit core: [1] and JGit: [2] and more or less inspiration for the project maintainer to have a look and decide whether or not this is useful.

* [1] https://gerrit-review.googlesource.com/#/c/gerrit/+/157073
* [2] https://git.eclipse.org/r/#/c/116659

Patch-set: 3
Reviewer: Gerrit User 1011323 <1011323@173816e5-2b9a-37c3-8a2e-48639d4f1153>