Add message for Empty Repositories
The empty repository experience can be confusing, as there is nothing rendered at all. This leads users to wonder if there was a problem rendering or if the repository is truly empty. This change adds a message that is shown the user in the event that there are no log, branches, nor tags in the repository.

Change-Id: I204b5f6090a7750d2ca89e5c36c1661fc3a3ddba