Fix empty lines not being copied to clipboard
Soy removes the empty (or useless) lines from the HTML code. When an
HTML element is open and directly closed, without any space or line
break, copying it results in nothingness. Before this change, when
copying a code block, the empty lines (from the code) would be missing.

The fix is to force an empty line when the line itself is empty. We
don't use a space as this would motify the copied code (the originally
empty line would now contain exactly one space).

Change-Id: I75fa24f47f91a6a553f3e982c0f1212d393c1443