Return unauthorized when repository cannot be accessed
Returning unauthorized instead for resources user do not have access to is
better indicator of what might be the reason of not finding the resource.

The actual benefit is that the gitiles plugin of gerrit detects this and
redirect the user to the login dialog.

This solves the "Not Found" white screen user get as greeting when people
send links to gitiles, while users do not have any feedback that login
should be performed nor there is a link to login.

Signed-off-by: Alon Bar-Lev <alon.barlev@gmail.com>
Change-Id: I47129cdbf2305298e437cb30ca0d97c2f9423b60
Bug: https://github.com/google/gitiles/issues/145