When resource is not found present decent page
This solves the "white screen" user get as greeting when people send links to
gitiles, while users do not have any feedback that login should be performed
nor there is a link to login.

The disadvantage of this method is that the gerrit plugin cannot determine
that it requires to redirect into login page, and the login page does not
redirect back to the page that not found. It requires much significant work in
gitiles gerrit plugin to have a decent user experience.

Signed-off-by: Alon Bar-Lev <alon.barlev@gmail.com>
Change-Id: Id3854ecc4fbe54e6a69112b0419edd29e8001694
Bug: https://github.com/google/gitiles/issues/145