Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> That's in response to ServiceNotAuthorizedException. What am I missing?

Look at all the examples and see that if *repository* is not found you get access denied and when *file* within a repository is found you get not found. This means that user experience with gitiles gerrit plugin is ok when sending a link to a repository but not for a file without the repository.

I have a different much more complex fix to show a proper page for not found within the view filter, but then realize that repository already return access denied and the gitiles gerrit plugin expects that... and this trivial patch solve it nicely with consistent behavior for both repository and repository file.

Patch-set: 3
Reviewer: Gerrit User 1040243 <1040243@173816e5-2b9a-37c3-8a2e-48639d4f1153>