Update patch set 3
Patch Set 3:

@Jonathan: before going into this direction, please make sure we understand how the menu item that gitiles gerrit plugin adds[1] can redirect back to the page.

This work seamlessly when returning access denied via [2] as the plugin is the one that handles the login.

While the menu items are plain urls that are common to the entire gitiles application. We can probably have a @CURRENT_URL_PLACEHOLDER@ to be used within registered url, as I am unsure we want to use referrer.

Expected user experience in this case:

1. User receives a link to a file in gitiles.
2. User opens the link and receives not found form.
3. User clicks sign in at the top left.
4. User is redirected to gerrit enters his credentials.
6. User is redirected back to the page and see the content.

[1] https://gerrit.googlesource.com/plugins/gitiles/+/master/src/main/java/com/googlesource/gerrit/plugins/gitiles/MenuFilter.java#44
[2] https://gerrit.googlesource.com/plugins/gitiles/+/master/src/main/java/com/googlesource/gerrit/plugins/gitiles/LoginFilter.java#44

Patch-set: 3
Reviewer: Gerrit User 1040243 <1040243@173816e5-2b9a-37c3-8a2e-48639d4f1153>