RefServlet: Use full reference in link
The servlet is checking for each reference if there is any other with
similar name. The only apparent benefit is to produce shorter links
(e.g. "test/+/foo" instead of "test/+/refs/heads/foo" when there is no
ambiguity on "foo").

This ambiguity check is an expensive operation. Use always full
reference name in the links. It is barely noticeable to the user and
speeds up the page generation.

Note that the short links are still valid, they are just not shown in
this page.

Change-Id: I1b6a15b3f584e85c56c08e31a24cdddc3129d883