Update patch set 4
Patch Set 4:

Hi,

Please understand that this is critical for people with private repositories (these that cannot be accessed by anonymous), this change solves the user experience nicely with a trivial change and with the current gitiles plugin, I urge you to consider merging this for now. We can then evaluate the #213695 change which requires probably additional changes to gitiles to allow menu items to receive the current resource to allow gitiles plugin to inject a login link that will return to the same resource.

We currently receive ACCESS DENIED when accessing a repository that is private, but not link, so I do not understand why this patch which aligns the behavior cannot be merged while considering a different unified solution.

People should be able to send links to files within repository browser is, it is a day to day activity, now that gitiles becomes part of gerrit, it must provide the expected user experience and interact with user to present login instead of blank page and after login to present the requested resource.

Thanks!

Patch-set: 4
Reviewer: Gerrit User 1040243 <1040243@173816e5-2b9a-37c3-8a2e-48639d4f1153>