Throw RequestFailureException instead of setting an error response code
There are cases where an exception is simply converted to an error
response. For example, with the following code the exception information
is lost and never logged in the server or shown to the client.

    try {
      // ...
    } catch (MissingObjectException | IncorrectObjectTypeException e) {
      res.setStatus(SC_NOT_FOUND);
      return;
    }

This makes it hard to debug a server bug. This change creates a new
RuntimeException class that indicates a request should be failed.

    try {
    } catch (MissingObjectException e) {
      throw new RequestFailureException(FailureReason.OBJECT_NOT_FOUND, e);
    } catch (IncorrectObjectTypeException e) {
      throw new RequestFailureException(FailureReason.INCORRECT_OBJECT_TYPE, e);
    }

This exception is handled in DefaultErrorHandlingFilter. This filter
can be replaced so that a server can change this implementation. This
gives a chance to handle errors nicely to the server. For example, it
can show the stack trace to the clients.

Change-Id: Ibb309121a7e152e65cb69c9ef9f8d8117aa0c4b2