Update patch set 1
Patch Set 1: Code-Review-1

Can this be split into two separate commits? First one to introduce and handle RequestFailureException, second to throw it in various contexts.

That way that first change would be less overwhelming to review (because it would be smaller) and can get the attention it deserves, and the second change would be less overwhelming to review (because it's essentially mechanical) and can get the attention it deserves.

Patch-set: 1
Reviewer: Gerrit User 1011842 <1011842@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Code-Review=-1