VisibilityCache: Use RefDatabase.getTipsWithSha1() for visibility check
The visibility check retrieves all refs, peels them and compares if
they point to the requested object. This is offered now in in the
RefDatabase.getTipsWithSha1() method. Using that method, we also
benefit from fine-tuned implementations for the specific storage in use.

Replace the loop with that method.

As a side, replace also the manual filter of the refs with the
RefDatabase.getRefsByPrefix(). This should delay as much as possible to
read all references (although real benefit depends on the
implemention).

This change requires https://git.eclipse.org/r/135547 in JGit,
available in JGit >= 5.4

Change-Id: I2623a3fc99cffcd791aa0c6450aba08d00408ca4
Signed-off-by: Ivan Frade <ifrade@google.com>