tools/eclipse/project.sh: Use bazel query to find project.py
This script generates the configuration for the eclipse editor running
the "project.py" script (provided by a bazlet).

To find this "project.py", the script uses a bazel workaround that seems
to be failing.

https://github.com/bazelbuild/bazel/issues/2452#issuecomment-316895752
reports an alternative to the workaround above, using bazel query.

Remove the py_binary workaround and use bazel query directly.

Validate also that the script is called from gitiles root directory.

Bug: Issue 10856
Change-Id: I91e47c2b1b167b38531f53a491683a3d47db9e00