Update patch set 10
Patch Set 10:

> I don't think anyone has time to work on this, and it's certainly no going to get into the stable-3.0 of Gerrit.  Getting it off the dashboard for now.

This is disappointing, it is once of the major gaps of gitiles over any other git browsing tools, including gitlab, github, gitweb and others. We git gitweb on gerrit only for that purpose, while it makes user experience very bad, people switch to getweb when required to download files.

Patch-set: 10