Update patch set 10
Patch Set 10:

> Patch Set 10:
> 
> > I don't think anyone has time to work on this, and it's certainly no going to get into the stable-3.0 of Gerrit.  Getting it off the dashboard for now.
> 
> This is disappointing, it is once of the major gaps of gitiles over any other git browsing tools, including gitlab, github, gitweb and others. We git gitweb on gerrit only for that purpose, while it makes user experience very bad, people switch to getweb when required to download files.

As disappointing as this is, there's not much can be done.  The original author of this change sadly passed away 18 months ago, the original developer of gitiles has moved on to another project, and the other contributors are doing so on a best effort basis.

If someone else wants to pick this up and work on it, I don't think anyone will object, but as Dave said in the earlier comment "this is a huge task".

Patch-set: 10