Update patch set 6
Patch Set 6:

(1 comment)

This requires[1].

Limitation: post login the gitiles plugin will redirect now to the home page. In any case this is required if [2] is not accepted as MVP which is how the gitiles plugin is designed to work.


[1] https://gerrit-review.googlesource.com/c/plugins/gitiles/+/211313
[2] https://gerrit-review.googlesource.com/c/gitiles/+/211312

Patch-set: 6
Reviewer: Gerrit User 1040243 <1040243@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Work-in-progress: false