Update patch set 9
Patch Set 9:

(1 comment)

Does this work for gitiles standalone?

Also, if the plugin changes strictly depend on this we cannot tie them together with the same topic. We will first need to build and release a new version of gitiles, with this change, and update core Gerrit to use it before we submit the plugin changes.

Patch-set: 9
Reviewer: Gerrit User 1011123 <1011123@173816e5-2b9a-37c3-8a2e-48639d4f1153>